requirements.txt
absl-py==0.8.1
altgraph==0.16.1
asgiref==3.2.3
asn1crypto==0.24.0
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
cachetools==3.1.1
cchardet==2.1.5
certifi==2019.11.28
chardet==3.0.4
convertapi==1.1.0
cycler==0.10.0
defusedxml==0.6.0
Django==3.0.1
django-allauth==0.41.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
et-xmlfile==1.0.1
extract-social-media==0.4.0
Flask==1.1.1
gast==0.3.2
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-translate==1.6.0
google-cloud-vision==0.39.0
google-pasta==0.1.7
googleapis-common-protos==1.6.0
grpcio==1.24.1
h5py==2.10.0
html-to-etree==0.2.0
httplib2==0.13.1
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
jdcal==1.4
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
minecart==0.3.0
more-itertools==7.2.0
oauth2client==4.1.3
oauthlib==3.1.0
opencv-contrib-python==4.0.0.21
opencv-python==3.4.4.19
openpyxl==2.5.12
packaging==19.2
pdf2image==1.8.0
pdfminer3k==1.3.1
pluggy==0.13.0
ply==3.11
protobuf==3.10.0
py==1.8.0
PyInstaller==3.5
pyPdf==1.13
PyPDF2==1.26.0
PyQt5==5.13.1
PyQt5-sip==4.19.19
PyQtWebEngine==5.13.1
pytest==5.2.0
python-axolotl==0.1.42
python-axolotl-curve25519==0.4.1.post2
python3-openid==3.1.0
pytz==2019.3
pywhatsapp==0.0.7
redis==3.3.8
requests==2.22.0
requests-file==1.4.3
requests-oauthlib==1.3.0
retrying==1.3.3
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
selenium==3.141.0
six==1.13.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tldextract==2.2.2
uritemplate==3.0.0
urllib3==1.25.7
w3lib==1.21.0
Wand==0.5.7
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
yowsup2==2.5.0
zipp==0.6.0
